Change return type of getStudentsQuantityByNames to Long

+++ src/main/java/ait/cohort46/student/controller/StudentController.java
@@ -50,7 +50,7 @@ public class StudentController {
     }
 
     @PostMapping("/quantity/students")
-    public Integer getStudentsQuantityByNames(@RequestBody Set<String> names) {
+    public Long getStudentsQuantityByNames(@RequestBody Set<String> names) {
         return studentService.getStudentsQuantityByNames(names);
     }
